% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{format_percent}
\alias{format_percent}
\title{Formatters}
\usage{
format_percent(number, decimal_places = 1)
}
\arguments{
\item{number}{number to turn into a percentage string, e.g. 10 to 10.0\% (numeric)}

\item{decimal_places}{number of decimal places to give percentage to (numeric)}
}
\value{
percentage formatted string (character)
}
\description{
Formatters
}
\examples{
format_percent(5, decimal_places = 2) # returns 5\%
}
