% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{format_scientific}
\alias{format_scientific}
\title{Formatters}
\usage{
format_scientific(number)
}
\arguments{
\item{number}{number to convert into scientific notation}
}
\value{
scientific notation form(character)
}
\description{
Formatters
}
\examples{
format_scientific(0.1)
}
