% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_barplot_counts}
\alias{geom_barplot_counts}
\title{geom barplot counts}
\usage{
geom_barplot_counts(
  distance_from_bar = 1.5,
  orientation = "h",
  size = 4,
  fontface = "bold",
  alpha = 0.8,
  color = "black",
  family = "Helvetica"
)
}
\arguments{
\item{distance_from_bar}{distance between text and cbar}

\item{orientation}{orientation of barplot ("h" / "horizontal" / "v" / "vertical")}

\item{size}{size of text (number)}

\item{fontface}{Font face ("plain", "italic", "bold", "bold.italic") (string)}

\item{alpha}{transparancy (number)}

\item{color}{colour (string)}

\item{family}{font family (string)}
}
\value{
ggplot geom
}
\description{
Add text labels indicating counts above barplot columns.
Works on plots that have a geom_bar() layer.
Works even if you flip axis with coord_flip().
}
\examples{
mtcars \%>\%
ggplot2::ggplot(ggplot2::aes(x=as.character(cyl))) +
ggplot2::geom_bar() +
ggplot2::xlab("cylinders") +
geom_barplot_counts()
}
