% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{geom_crossbar_predefined}
\alias{geom_crossbar_predefined}
\title{Add crossbar to ggplot}
\usage{
geom_crossbar_predefined(
  summaryfunction = stats::median,
  width = 0.4,
  size = 0.3,
  colour
)
}
\arguments{
\item{summaryfunction}{a function run on the y aesthetic to determine where line is drawn.
Options include median, mean, max, min, or any other function that summarises a numeric vector into a single number (function)}

\item{width}{width of crossbar}

\item{size, colour}{ggplot aesthetics}
}
\value{
ggplot geom
}
\description{
Adds a crossbar to a ggplot.
Best used when comparing one categorical and one numeric variable using geom_point / geom_jitter(height=0).
}
\examples{
mtcars \%>\%
  ggplot2::ggplot(ggplot2::aes(cyl>6, mpg)) +
  ggplot2::geom_point() +
  geom_crossbar_predefined()
}
