% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{ggbarplot}
\alias{ggbarplot}
\title{Plot}
\usage{
ggbarplot(
  data,
  col,
  orientation = "h",
  fill = NULL,
  title = NULL,
  position = "stack",
  ...
)
}
\arguments{
\item{data}{a dataset to plot (dataframe)}

\item{col}{column of dataframe to plot (don't quote)}

\item{orientation}{orientation of barplot ("h" / "horizontal" / "v" / "vertical")}

\item{fill}{column of dataframe to color based on (don't quote)}

\item{title}{title of graph (string)}

\item{position}{see ?ggplot2::geom_bar for details (Usually one of: "stack", "dodge", "fill")}

\item{...}{set any other barplot property. See ?ggplot2::geom_bar for details. (e.g. alpha = 0.4)}
}
\value{

}
\description{
Plot
}
\examples{
ggbarplot(iris, Species, fill = Species, orientation = "horizontal", title = "My Freq Graph")
}
