% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R.R
\name{ggdensity}
\alias{ggdensity}
\title{Density Plot}
\usage{
ggdensity(
  data,
  col,
  fill = "steelblue",
  alpha = 0.5,
  summary_stats = FALSE,
  text_xpos = NULL,
  text_ypos = 0.04,
  text_sigfigs = Inf,
  ...
)
}
\arguments{
\item{data}{datafrmae}

\item{col}{column (don't quote)}

\item{fill}{geom_density fill color}

\item{alpha}{geom_density alpha color}

\item{summary_stats}{add text with summary stats to plot  (boolean)}

\item{text_xpos}{x position of summary stat text (numeric)}

\item{text_ypos}{y position of summary stat text (numeric)}

\item{text_sigfigs}{number of significant figures to write summary statistic to (numeric)}

\item{...}{other geom_density aesthetics (e.g. linetype = "dashed")}
}
\value{
gglot
}
\description{
Density Plot
}
\examples{
ggdensity(mtcars, mpg, summary = TRUE)
}
