% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_common_adjustments}
\alias{theme_common_adjustments}
\title{Custom Themes}
\usage{
theme_common_adjustments(
  dist_from_plot_xlab = 10,
  dist_from_plot_ylab = 10,
  dist_from_plot_ggtitle = 10,
  no_background = FALSE,
  subtitle_face = "plain"
)
}
\arguments{
\item{dist_from_plot_xlab}{distance between x axis title and plot (number)}

\item{dist_from_plot_ylab}{distance between y axis title and plot (number)}

\item{dist_from_plot_ggtitle}{distance between y axis title and plot (number)}

\item{no_background}{set all backgrounds to clear? (flag)}

\item{subtitle_face}{Font face ("plain", "italic", "bold", "bold.italic")}
}
\value{
ggtheme
}
\description{
ggplot theme that fixes a bunch of issues I tend to have with plots.
}
\details{
By default, will bold center axis and plot titles and tweak distane of axis_titles to plot
}
\examples{
mtcars \%>\%
  ggplot2::ggplot(ggplot2::aes(cyl>6, mpg)) +
  ggplot2::geom_point() +
  theme_common_adjustments()
}
