% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_legend_bottom}
\alias{theme_legend_bottom}
\title{Custom Themes}
\usage{
theme_legend_bottom(direction = "horizontal")
}
\arguments{
\item{direction}{"Vertical or Horizontal"}
}
\value{
ggtheme
}
\description{
Custom Themes
}
