% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_minimal_bordered}
\alias{theme_minimal_bordered}
\title{Miminal theme with border}
\usage{
theme_minimal_bordered(border_color = "grey40", border_thickness = NULL, ...)
}
\arguments{
\item{border_color}{Colour of border}

\item{border_thickness}{border thickness in mm}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal}}
  \describe{
    \item{\code{base_size}}{base font size, given in pts.}
    \item{\code{base_family}}{base font family}
    \item{\code{base_line_size}}{base size for line elements}
    \item{\code{base_rect_size}}{base size for rect elements}
  }}
}
\description{
A minimal theme with a border. Works well for faceted graphs
}
\examples{
mtcars \%>\%
  ggplot2::ggplot(ggplot2::aes(x = mpg, y=disp)) +
  ggplot2::geom_point() +
  theme_minimal_bordered()
}
