% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_no_legend}
\alias{theme_no_legend}
\title{Remove theme legend}
\usage{
theme_no_legend(...)
}
\arguments{
\item{...}{no arguments are used. Included only so code written for older versions of package doesn't break}
}
\value{
theme
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
