% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_show_infinites_y}
\alias{scale_show_infinites_y}
\title{Include Infinite Values in Plot}
\usage{
scale_show_infinites_y()
}
\value{
ScaleContinuousPosition object that can be added to a ggplot object using '+'
}
\description{
Will replace infinite values by the nearest limit
}
\examples{
mtcars \%>\%
    dplyr::mutate(qsec2 = ifelse(qsec > 19, Inf, qsec)) \%>\%
    ggplot2::ggplot(ggplot2::aes(x=qsec2, y=carb)) +
    ggplot2::geom_point() +
    scale_show_infinites_y()
}
\seealso{
\code{\link[=scale_show_infinites_x]{scale_show_infinites_x()}}
}
